backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
redis
celery
sentence-transformers
spacy
pydantic
python-multipart
pandas
numpy
python-dotenv
alembic
pydantic-settings
google-genai
